pyproject.toml
[project]
name = "llmsurgeon"
version = "0.1.0"
description = "LLMSurgeon: Diagnosing Data Mixture of Large Language Models"
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Yaxin Luo" },
    { name = "Jiacheng Cui" },
    { name = "Xiaohan Zhao" },
    { name = "Xinyi Shang" },
    { name = "Jiacheng Liu" },
    { name = "Xinyue Bi" },
    { name = "Zhaoyi Li" },
    { name = "Zhiqiang Shen" },
]
keywords = [
    "language-models",
    "data-mixture",
    "label-shift",
    "membership-inference",
    "pretraining-data",
    "acl2026",
]
dependencies = [
    "torch",
    "transformers",
    "scikit-learn",
    "numpy",
    "pandas",
    "tqdm",
    "matplotlib",
    "pyyaml",
    "datasets",
    "accelerate",
    "seaborn",
    "tabulate",
    "zstandard",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.0.250",
]

[project.urls]
# TODO: fill in before public release
Homepage = "https://github.com/<org>/LLMSurgeon"
Repository = "https://github.com/<org>/LLMSurgeon"
Paper = "https://arxiv.org/abs/TBD"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["baseline_method/src/labelshift"]

